refactor(completion-records): break queue storage cycle
